requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
playwright>=1.30.0
pyyaml>=6.0
